Fix/mini auth

+++ mom-auth-platform/mom-auth-server/src/main/java/io/github/chrisshi/mom/auth/application/RoleApplication.java
@@ -303,7 +303,7 @@ public class RoleApplication {
     }
 
     private static void requireVersion(Long actual, long expected) {
-        if (actual == null || actual.longValue() != expected) {
+        if (actual == null || actual != expected) {
             throw new AuthException(AuthErrorCode.OPTIMISTIC_LOCK_CONFLICT);
         }
     }
